refactor: dedupe MCP/LSP tool expr builders into beamtalk-core BT-3193
save_class_expr, precheck_method_expr, remove_method_expr,
remove_method_if_absent_expr, and flush_expr/FlushFilter were
hand-rolled independently in beamtalk-mcp and beamtalk-lsp, documented
only in prose as "mirroring" each other with no test enforcing it. Per
architecture-principles.md's shared-leaf-module pattern, extract them
into beamtalk_core::tool_expr (both crates already depend on
beamtalk-core) so the two surfaces call one implementation instead of
two that could drift. The module's unit tests are the BT-3193
conformance suite; a single definition can't disagree with itself.
