+++ crates/beamtalk-core/src/lib.rs
@@ -33,6 +33,7 @@ pub mod source_analysis;
 pub(crate) mod state_threading_selectors;
 pub(crate) mod synthetic_selectors;
 pub mod test_helpers;
+pub mod tool_expr;
 pub mod unparse;
 
 /// Re-export commonly used types.
